Update projections.tex
Added a note on epsilon in Theorem 7 being the parameter of the distribution and not the accuracy. Also, added a short discussion on the simple analysis of Theorem 7 when the learner is proper. Finally, edited some of the formulas in the displays to get rid of the overfull boxes (some of which lead to failure when running the ICML paper checker, some just overlapped with the text in the other column).